Have client views and search work. Mostly.
This has a recycler view that displays the clients. For now, instead
of the real scheme, this takes the first char of the filename for
the name and the second for the reading.
This implements search in a rudimentary way : it waits for 1500 ms
after char input then performs the search. This needs to be improved
with cancellation and immediate search when available.
One of the views is truncated in width, and I don't know why. I
suspect it's laid out with a first string ("i" is a good candidate,
or "f" maybe) and not laid out again later.