Implement save to cloud and restore.
This adds icons for save and undo. Undo is not wired yet and does
nothing at this time.
This stores the drawing as an encoded list of floats out of which
the drawing can be replayed. For security, the replay is wired
through the same functions as the actual drawing ; because each
function is storing the data itself it's not strictly speaking a
guarantee it will be drawn the same, but it's still very easy to
get right.
While the drawing is entirely done in a normalized coordinate system
to ensure it looks the same no matter the resolution of the device,
the rendering is still done natively to the device resolution to
ensure the best possible quality on any given device. The picture
however is exported to the cloud in native resolution (we may want
to change this in the future), which is simple and easy and reasonably
intuitive, but it does mean the exported picture's resolution depends
on the device's resolution. I think it's fine, but if it's a problem
later it should not be very hard to change – any device can render
at any resolution if that's what we want.

At this commit though there is a small bug in the export : the
guide picture is blitted on the exported image fitting both width
and height, ignoring the aspect ratio, so it's distorted on the final
renders.