Straighten out the save lifecycle.
Back button or change drawing will save the drawing immediately.
Any action will enqueue a 4 second message to save.

This is not very pleasing, and there may be race conditions.
But for now it will do. It's probably fine in 99% cases.

Also, whenever the canvas is touched, remove the 'saved' icon.
This does not work for the back button because by then the
interface has gone back to the list, but I think that's actually
better not to remove the icon in this case.