Switch to the REST API.
It's not as good, but at least it's not deprecated.
It doesn't do caching, which is an extreme bother to reimplement :(
AFAICT this is fully working now with the exception of caching,
and a bug where going back to the client list after creating a
new client or updating its metadata is not reflected in the list
of clients.