Rename the png files to .png
This will let Drive understand the preview. It kind of sucks that
it seems it's the only way, but if you think of it, when you
download the files, poor OSes with poor code will not be able to
open the files easily if they don't have an extension, so might
as well put it.