Implement most of the load-after-cache logic.
Sometimes data is cached and the network is slow. The app
will use the cache, but after a while the network access
returns.
When that happens, compare loaded data to the cached data
that has been used. If they match, do nothing, but if they
don't, warn the caller.

A Flow is used to do that, to force the caller to take
this possibility into account ; they have to collect the
flow which forces them to think it may happen multiple
times.

This patch also implements handling of that for the
session data, in two instances :
• In the session data list, where it should just be
  displayed again.
• In the session editor, which now will display a message
  prompting the user to keep the current changes and
  overwrite the updated data, or throw out the current
  changes and use the updated data. If there were no local
  changes to the data since the data was opened, (but
  there still was updated data from the network, otherwise
  this flow won't trigger at all of course), then don't
  even give the option to "keep current changes", force
  opening updated data.