Fix : updates to the data not reflected up
When going back to the list of sessions the data just
edited was not reflected. The reason is that Network
data was prioritized always, and LiveCache just hangs
on foreven to the stuff it fetches once, even if it's
edited locally. The same bug exists for lists, but it
will be fixed in subsequent patches.