Fix bugs
The big one was an issue where doing anything in Drive
would require the root folder of the app, which would
wait on a command, that would wait on the full list of
outstanding commands. So if the DB still contains a
command to put a file, that blocks on creating the
root folder of the app, which waits for a command that's
after it in the DB. Note that this means you need
initial internet access to start the app, which is
not really acceptable after the first run, but it was
already the case in the old code.

Another one was the creation of a separate Root for
UI and background commands, where the app should use
the same for both when possible. This raises the
question of whether uploading in the background works
at all... Likely, the problem is getting permisssions.

Yet another one was another bug in the way the code
loops over the listeners. Turns out Kotlin is awful at
doing this, because forEach uses an iterator so you
can't remove own stuff, and the natural way of reverse
iterating is to use a for loop with an `in` range
starting at -1... which means the range can take the
unexpected value -1..0 and throw OOBE.